General Bug fix and file uploading

+++ cloudcw/src/main/java/com/pg4/cloudcw/service/FolderService.java
@@ -7,8 +7,13 @@ import javax.transaction.Transactional;
 
 import org.springframework.stereotype.Service;
 
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
+import org.joda.time.format.DateTimeFormat;
+
 import com.pg4.cloudcw.dao.FolderRepository;
 import com.pg4.cloudcw.entity.File;
+import com.pg4.cloudcw.entity.Flag;
 import com.pg4.cloudcw.entity.Folder;
 
 @Service
@@ -21,10 +26,24 @@ public class FolderService {
 		this.folderRepository = folderRepository;
 	}
 
+	// Get a folder By Id
+	public Folder getFolderById(int id) {
+		return folderRepository.findById(id);
+	}
+	
 	// Get all Folders By User Id
 	public List<Folder> getAllByUserId(int userId) {
 		List<Folder> folders = new ArrayList<>();
 		for (Folder f : folderRepository.findByUserIdAndIsDeleted(userId, false)) {
+			folders.add(f);
+		}
+		return folders;
+	}
+	
+	// Get all Folders By User and Folder
+	public List<Folder> getAllByUserIdandFolder(int userId, int folderId) {
+		List<Folder> folders = new ArrayList<>();
+		for (Folder f : folderRepository.findByUserIdAndParentFolderIdAndIsDeleted(userId, folderId,false)) {
 			folders.add(f);
 		}
 		return folders;
@@ -40,18 +59,31 @@ public class FolderService {
 	}
 
 	public void createFolder(Folder newFolder) {
+		if(folderRepository.existsByName(newFolder.getName())){
+			String uniqueName = DateTime.now(DateTimeZone.UTC).toString(DateTimeFormat.forPattern("-YYYY-MM-dd-HHmmss"));
+			newFolder.setName(newFolder.getName()+ uniqueName);
+		}
 		folderRepository.save(newFolder);
 	}
 	
-	// For Trash folder
-	public void renameFolder(int id, int userId, String NewName) {
+	// Rename folder
+	public void renameFolder(int id, int userId, String newName) {
 		Folder folder = folderRepository.findById(id);
 		if (userId == folder.getUser().getId()) {
-			folder.setName(NewName);
+			folder.setName(newName);
 			folderRepository.save(folder);
 		}
 	}
-
+	
+	// Change folder flag
+	public void changeFolderFlag(int id, int userId, Flag newFlag) {
+		Folder folder = folderRepository.findById(id);
+		if (userId == folder.getUser().getId()) {
+			folder.setFlag(newFlag);
+			folderRepository.save(folder);
+		}
+	}
+	
 	// For changing Current Directory
 	public void changeFolderDirectory(int id, int userId, int parentFolderId) {
 		Folder folder = folderRepository.findById(id);
@@ -60,7 +92,6 @@ public class FolderService {
 			folderRepository.save(folder);
 		}
 	}
-	
 	
 	// For Trash folder
 	public void deleteFirstTime(int id, int userId) {
@@ -81,7 +112,17 @@ public class FolderService {
 	}
 
 	//Delete Permanently from Trash
-	public void deletePermanently(Folder folder) {
-		folderRepository.delete(folder);
+	public void deletePermanently(int folderId) {
+		folderRepository.deleteById(folderId);
+	}
+	
+	// Delete All Files Permanently from Trash
+	public void deleteAllPermanently(int userId) {
+		List<Folder> folderList = folderRepository.findByUserIdAndIsDeleted(userId, true);
+		for (Folder folder : folderList) { // TODO: User Check
+			if (folder.getUser().getId() == userId) {
+				folderRepository.delete(folder);
+			}
+		}
 	}
 }
